Fix module searchbar, still methods missing
