+++ src/App.js
@@ -1,5 +1,7 @@
 import "./app.css";
+
 import { createElement } from "./utils/elements";
+
 import { getCharactersByName } from "./utils/api";
 
 import { Character } from "./components/Character";
@@ -12,32 +14,32 @@ function App() {
   const header = Header();
   page.append(header);
 
-  const searchBar =  Searchbar();
-  page.append(searchbar);
+
+  const searchBar = Searchbar();
+  searchBar.oninput = (event) => {
+    const characterContainer = document.querySelector(".main");
+    while (characterContainer.firstChild) {
+      characterContainer.removeChild(characterContainer.firstChild);
+    }
+    event.preventDefault();
+    searchCharacters(document.querySelector("#input").value);
+  };
+
+  searchBar.onsubmit = (event) => {
+    const characterContainer = document.querySelector(".main");
+    while (characterContainer.firstChild) {
+      characterContainer.removeChild(characterContainer.firstChild);
+    }
+    event.preventDefault();
+    searchCharacters(document.querySelector("#input").value);
+    document.querySelector("#input").value = "";
+  }
+  page.append(searchBar);
 
   const main = createElement("article", { className: "main" });
   page.append(main);
 
- /*
- * Get one page after another
-
-  async function getMoreCharacter(page = 1) {
-    const characters = await getCharactersByPage(page);
-    // if lesser are needed: const newArray = characters.slice(startIndex, amountItems);
-    const characterElements = characters.map((singleCharacter) => {
-      return Character({
-        name: singleCharacter.name,
-        imgSrc: singleCharacter.image,
-      });
-    });
-    main.append(...characterElements);
-  } 
-  
-  getMoreCharacter();
-  
-  */
-
- export async function searchCharacters(query) {
+  async function searchCharacters(query) {
     const characters = await getCharactersByName(query);
     // if lesser are needed: const newArray = characters.slice(startIndex, amountItems);
     const characterElements = characters.map((singleCharacter) => {
